requirements.txt
# --- Core ---
Flask==2.3.3
mysql-connector-python==9.1.0
python-dotenv==1.0.1
requests==2.31.0

# --- Parsing / HTML ---
beautifulsoup4==4.12.2
lxml==4.9.3
chardet==5.2.0
regex

# --- Google Gemini ---
google-generativeai==0.7.2

# --- Email handling ---
gunicorn==21.2.0

# --- Optional NLP / Logging ---
nltk==3.8.1
